% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.slca}
\alias{confint.slca}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{slca}(object, parm, level = 0.95, type = c("param", "logit"), ...)
}
\arguments{
\item{object}{an object of class \code{slca} and \code{estimated}.}

\item{parm}{an integer string specifying parameters to compute confidence intervals.}

\item{level}{level a numeric value representing the desired confidence level for the intervals. Default is 0.95}

\item{type}{a character string specifying the format in which the results should be returned. Options are \code{"probs"} for probability format and \code{"logit"} for log-odds (logit) format. The default is \code{"probs"}.}

\item{...}{additional arguments.}
}
\description{
Computes confidence intervals for one or more parameters of fitted model. Package \pkg{slca} adds methods for \code{slca} fits.
}
