% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{param}
\alias{param}
\alias{param.slca}
\title{Printing Estimated Parameters of \code{slca} Object}
\usage{
param(object, ...)

\method{param}{slca}(
   object, type = c("probs", "logit"),
   se = FALSE, index = FALSE, ...
)
}
\arguments{
\item{object}{an object of class \code{slca} and \code{estimated}.}

\item{...}{additional arguments.}

\item{type}{a character string indicating the format in which the estimated parameters should be returned. If set to \code{"probs"}, estimates are returned in probability form, while \code{"logit"} returns them in log-odds (logit) form. Default is \code{"probs"}.}

\item{se}{a logical value whether to print parameter estimates or standard errors.}

\item{index}{a logical value. If \code{TRUE}, the indices of the estimated parameters are included in the output, enclosed in parentheses. If \code{FALSE}, they are omitted.}
}
\value{
A list containing the specified estimated parameters (or corresponding standard errors for \code{se} function):
\item{pi}{Membership probabilities of the root variable.}
\item{tau}{Conditional probabilities between latent class variables, represented with uppercase alphabets for considering measurement invariance.}
\item{rho}{Item response probabilities for each measurement model, represented with lowercase alphabets for considering measurement invariance.}
}
\description{
By passing \code{estimated} \code{slca} object, the function prints estimated parameters of the slca model.
}
