% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slca-package.R
\docType{package}
\name{slca-package}
\alias{slca-package}
\title{slca: An R Package for Structural Modeling for Multiple Latent Class Variables}
\description{
\code{slca} package provides tools for the easy implementation of Structural Latent Class Models (SLCM), including Latent Transition Analysis (LTA), Latent Class Profile Analysis (LCPA), and Joint Latent Class Analysis (JLCA).
}
\author{
\strong{Maintainer}: Youngsun Kim \email{yskstat@gmail.com} (\href{https://orcid.org/0000-0001-8003-1939}{ORCID})

Authors:
\itemize{
  \item Hwan Chung \email{hwanch@korea.ac.kr} (\href{https://orcid.org/0000-0002-8969-9086}{ORCID})
}

}
\keyword{internal}
