% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsy_fit.R
\name{nlsy_jlcpa}
\alias{nlsy_jlcpa}
\alias{nlsy_jlta}
\title{JLCPA Model Estimated with NLSY97 Data}
\format{
An \code{slca} object \code{estimated} for JLCPA model using \code{nlsy97}.
}
\description{
The \code{slca} model estimated using NLSY97 data
}
\references{
Bureau of Labor Statistics, U.S. Department of Labor. National Longitudinal Survey of Youth 1997 cohort, 1997-2017 (rounds 1-18). Produced and distributed by the Center for Human Resource Research (CHRR), The Ohio State University. Columbus, OH: 2019.

Jeon, S., Seo, T. S., Anthony, J. C., & Chung, H. (2022). Latent Class Analysis for Repeatedly Measured Multiple Latent Class Variables. Multivariate Behavioral Research, 57(2–3), 341–355.
}
\seealso{
\link[slca]{reorder.slca}
}
