% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.slca.R
\name{simulate.slca}
\alias{simulate.slca}
\title{Simulate data from \code{slca} model.
This function simulate data from constructed \code{slca} model.
If the object is not estimated, the parameters can be given by user or can be generated at random.}
\usage{
\method{simulate}{slca}(object, nsim = 500, seed = NULL, parm, nlevel, ...)
}
\arguments{
\item{object}{a \code{slca} object}

\item{nsim}{number of response observation to be simulated. Defaults to 500.}

\item{seed}{random seed.}

\item{parm}{parameter set by users.}

\item{nlevel}{number of levels for each manifest items declared in the model.}

\item{...}{additional arguments.}
}
\description{
Simulate data from \code{slca} model.
This function simulate data from constructed \code{slca} model.
If the object is not estimated, the parameters can be given by user or can be generated at random.
}
\examples{
library(magrittr)
model <- slca(lc1[3] ~ x1 + x2 + x3 + x4 + x5)
sim_data <- model \%>\% simulate(nlevel = c(3, 3, 3, 3, 3))
y <- sim_data$response
sapply(y, table)

sim_data <- model \%>\%
   simulate(nlevel = c(x1 = 2, x3 = 3, x4 = 4, x5 = 5))
y <- sim_data$response
sapply(y, table)

model <- slca(lc1[3] ~ x1 + x2 + x3 + x4 + x5,
              lc2[4] ~ y1 + y2 + y3 + y4 + y5)
sim_data <- model \%>\% simulate(1000)
sapply(sim_data$class, table)
}
