% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.slca}
\alias{simulate.slca}
\title{Simulate Data from \code{slca} Model}
\usage{
\method{simulate}{slca}(object, nsim = 500, seed = NULL, parm, nlevel, ...)
}
\arguments{
\item{object}{a \code{slca} object representing the model from which data will be simulated.}

\item{nsim}{the number of response observations to simulate. Defaults to 500.}

\item{seed}{a random seed for reproducibility of the}

\item{parm}{a set of parameters provided by the user to guide the simulation, if the model has not been estimated.}

\item{nlevel}{the number of levels for each manifest item declared in the model. If not provided, the default is 2.}

\item{...}{additional arguments.}
}
\value{
A \code{list} of two components:
\item{class}{A \code{data.frame} providing the assigned latent class for each individual across different latent class variables.}
\item{response}{A \code{data.frame} containing the manifest items that were simulated.}
}
\description{
This function simulates data from a constructed \code{slca} model. If the model is not already estimated, parameters can be specified by the user or generated randomly.
}
\examples{
m1 <- slca(lc1[3] ~ x1 + x2 + x3 + x4 + x5,
              lc2[4] ~ y1 + y2 + y3 + y4 + y5)
sim <- simulate(m1, 1000)
sapply(sim$class, table)

# simulate data with defined number of levels of manifest items
m2 <- slca(lc1[3] ~ x1 + x2 + x3 + x4)
sim <- simulate(m2, nlevel = c(3, 3, 3, 3))
d <- sim$response
sapply(d, table)

sim <- simulate(m2, nlevel = c(x1 = 2, x3 = 3, x4 = 4, x5 = 5))
d <- sim$response
sapply(d, table)

# simulate data with user-defined parameters
pi <- rep(1 / 3, 3)
rho <- c(.9, .1, .9, .1, .9, .1, .9, .1,
         .9, .1, .9, .1, .1, .9, .1, .9,
         .1, .9, .1, .9, .1, .9, .1, .9)
par <- c(pi, rho)
m3 <- slca(lc[3] ~ y1 + y2 + y3 + y4)
sim <- simulate(m3, parm = par)
mf <- estimate(m3, sim$response)
param(mf)
}
