% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slca-package.R
\docType{package}
\name{slca-package}
\alias{slca-package}
\title{slca: An R Package for Structural Modeling for Multiple Latent Class Variables}
\description{
The \code{slca} package provides a comprehensive toolkit for the implementation of Structural Latent Class Models (SLCM). This includes tools for conducting Latent Transition Analysis (LTA), Latent Class Profile Analysis (LCPA), and Joint Latent Class Analysis (JLCA), facilitating the analysis of complex latent structures in statistical data.
}
\author{
\strong{Maintainer}: Youngsun Kim \email{yskstat@gmail.com} (\href{https://orcid.org/0000-0001-8003-1939}{ORCID})

Authors:
\itemize{
  \item Hwan Chung \email{hwanch@korea.ac.kr} (\href{https://orcid.org/0000-0002-8969-9086}{ORCID})
}

}
\keyword{internal}
